requirements.txt
asttokens==2.0.5
backcall==0.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
dbus-python==1.2.18
decorator==5.1.1
docopt==0.6.2
executing==0.8.3
future==0.18.2
idna==3.3
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.3
numpy==1.23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pkg_resources==0.0.0
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.21.0
Pygments==2.12.0
PyGObject==3.42.2
pyserial==3.5
pyusb==1.2.1
requests==2.28.1
rfcat==1.9.5
sh==1.14.2
six==1.16.0
stack-data==0.3.0
termcolor==1.1.0
traitlets==5.3.0
urllib3==1.26.10
wcwidth==0.2.5
yarg==0.1.9
